+++ server/view/index.html
@@ -7,5 +7,5 @@
 </head>
 <body>
     <div id='root'></div>
-<script src="react.bundle.ddb7a817577dc5f18f43.js"></script><script src="vendors.ddb7a817577dc5f18f43.js"></script><script src="app.ddb7a817577dc5f18f43.js"></script></body>
+<script src="react.bundle.9605a2d491690cf6b90f.js"></script><script src="vendors.9605a2d491690cf6b90f.js"></script><script src="app.9605a2d491690cf6b90f.js"></script></body>
 </html>
