+++ src/gnb/component/LoginBtn.js
@@ -1,13 +1,9 @@
-import { useState, useEffect } from "react";
 import { Button, Typography, Divider, IconButton } from "@material-ui/core";
 import styled, { css } from "styled-components";
 import { Link } from "react-router-dom";
 import Modal from "../../common/component/Modal";
 import PhotoCamera from "@material-ui/icons/PhotoCamera";
 import SignIn from "./Signin";
-import axios from "axios";
-import { useSelector, useDispatch } from "react-redux";
-import { actions } from "../../common/reducer/user";
 
 const UserContainer = styled.div`
   display: flex;
@@ -78,36 +74,13 @@ const Avatar = styled(IconButton)`
     `}
 `;
 
-export default function Login() {
-  const [loginModal, setLoginModal] = useState(false);
-  const [infoModal, setInfoModal] = useState(false);
-  const [userInfo, setUserInfo] = useState(null); //ㄴㄴ
-
-  const dispatch = useDispatch();
-  const { isLoggedIn, info } = useSelector((state) => state.user);
-
-  useEffect(() => {
-    console.log("------BTN useEffect");
-    console.log(info);
-
-    if (isLoggedIn) return;
-    dispatch(actions.loginInRequest());
-  }, [dispatch, isLoggedIn]);
-
-  const handleClickLoginModal = () => {
-    setLoginModal(!loginModal);
-  };
-
-  const handleClickInfoModal = () => {
-    setInfoModal(!infoModal);
-  };
-
-  const handleAddFile = (e) => {
-    const formData = new FormData();
-    formData.append("img", e.target.files[0]);
-    dispatch(actions.uploadImageRequest(formData));
-  };
-
+export default function Login({
+  handleAddFile,
+  handleClickInfoModal,
+  handleClickLoginModal,
+  loginModal,
+  infoModal,
+}) {
   return (
     <>
       {isLoggedIn ? (
